+++ packages/client/mutations/AutogroupMutation.ts
@@ -7,15 +7,13 @@ graphql`
   fragment AutogroupMutation_meeting on AutogroupSuccess {
     meeting {
       id
-      resetReflectionGroups {
-        groupTitle
-      }
       reflectionGroups {
         id
         title
         reflections {
           id
           plaintextContent
+          reflectionGroupId
         }
       }
     }