+++ scripts/toolboxSrc/primeIntegrations.ts
@@ -1,3 +1,4 @@
+import getKysely from '../../packages/server/postgres/getKysely'
 import upsertIntegrationProvider from '../../packages/server/postgres/queries/upsertIntegrationProvider'
 import {Logger} from '../../packages/server/utils/Logger'
 
@@ -51,6 +52,17 @@ const upsertGlobalIntegrationProvidersFromEnv = async () => {
       (authStrategy === 'oauth2' && clientId && clientSecret && serverBaseUrl) ||
       (authStrategy === 'sharedSecret' && sharedSecret && serverBaseUrl)
   )
+
+  // it's a security risk to have a misconfigured mattermost provider
+  if (!process.env.MATTERMOST_URL || !process.env.MATTERMOST_SECRET) {
+    const pg = getKysely()
+    await pg
+      .deleteFrom('IntegrationProvider')
+      .where('service', '=', 'mattermost')
+      .where('authStrategy', '=', 'sharedSecret')
+      .where('scope', '=', 'global')
+      .execute()
+  }
 
   await Promise.all(
     validProviders.map((provider) => {